enhanced transport class to be able to make standard GET requests
changed wsdl class to be able to accept proxy info, and to use the transport class for retreiving wsdl files from network.

changed soapclient to be able to pass on it's proxy info to the wsdl class. this only works when the user has extended the soapclient class, and calls setHTTPPRoxy prior to calling the soapclient constructor (eventually there'll be a more elegant way of doing this).

built new nusoap.php